separate platform jest configs, uncomment android test
