+++ packages/expo-updates/e2e/jest.config.android.js
@@ -1,0 +1,6 @@
+const expoDefault = require('expo-module-scripts/jest-preset-plugin');
+
+module.exports = {
+  ...expoDefault,
+  moduleFileExtensions: ['android.ts', 'ts', 'js', 'json'],
+};
