+++ packages/expo-updates/e2e/jest.config.ios.js
@@ -1,0 +1,6 @@
+const expoDefault = require('expo-module-scripts/jest-preset-plugin');
+
+module.exports = {
+  ...expoDefault,
+  moduleFileExtensions: ['ios.ts', 'ts', 'js', 'json'],
+};
