--- packages/expo-updates/e2e/jest.config.js
@@ -1,1 +1,0 @@
-module.exports = require('expo-module-scripts/jest-preset-plugin');
